Extract decoding functionality to improve testing and modularity
